refactoring tabs into their own components
