+++ packages/web/src/components/nutrition/modal/recipeModal.tsx
@@ -1,12 +1,14 @@
 import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 import Modal from '@mui/material/Modal'
-import { RecipeModel } from '../../models/recipeModels'
-import NutritionFacts from './nutritionFacts/nutritionFacts'
+import { RecipeModel } from '../../../models/recipeModels'
+import NutritionFacts from '../nutritionFacts/nutritionFacts'
 import { useState } from 'react'
 import Tabs from '@mui/material/Tabs'
 import Tab from '@mui/material/Tab'
-import TabPanel from '../TabPanel'
+import TabPanel from '../../TabPanel'
+import RecipeModalOverview from './recipeModalOverview'
+import RecipeModalDirections from './recipeModalDirections'
 
 interface Props {
   open: boolean
@@ -64,26 +66,13 @@ export default function RecipeModal(props: Props) {
             </Tabs>
           </Box>
           <TabPanel value={tab} index={0}>
-            <Typography id="modal-modal-description" sx={{ mt: 2 }}>
-              {props.recipe.summary}
-            </Typography>
+            <RecipeModalOverview recipe={props.recipe} />
           </TabPanel>
           <TabPanel value={tab} index={1}>
             <NutritionFacts recipe={props.recipe} />
           </TabPanel>
           <TabPanel value={tab} index={2}>
-            <h4>Ingredients:</h4>
-            <ul>
-              {props.recipe.ingredients.map((ingredient) => {
-                return <li key={ingredient}>{ingredient}</li>
-              })}
-            </ul>
-            <h4>Directions:</h4>
-            <ul>
-              {props.recipe.directions.map((direction) => {
-                return <li key={direction}>{direction}</li>
-              })}
-            </ul>
+            <RecipeModalDirections recipe={props.recipe} />
           </TabPanel>
         </Box>
       </Box>
