+++ packages/web/src/components/nutrition/recipeSummary.tsx
@@ -1,6 +1,6 @@
 import { RecipeModel } from '../../models/recipeModels'
 import { useEffect, useState } from 'react'
-import RecipeModal from './recipeModal'
+import RecipeModal from './modal/recipeModal'
 import Pagination from '@mui/material/Pagination'
 import '../../App.css'
